+++ src/Common/JK.Common/TypeHelpers/GZipHelper.cs
@@ -1,0 +1,47 @@
+﻿using System.IO;
+using System.IO.Compression;
+
+namespace JK.Common.TypeHelpers
+{
+    /// <summary>
+    /// Class which contains methods for GZip compression.
+    /// </summary>
+    public class GZipHelper
+    {
+        /// <summary>Compresses a file using GZip compression.</summary>
+        /// <param name="inFileName">File name of original file.</param>
+        /// <param name="outFileName">File name to give to compressed file.</param>
+        public void GZipCompressFile(string inFileName, string outFileName)
+        {
+            using (FileStream sourceFile = File.OpenRead(inFileName))
+            using (FileStream destFile = File.Create(outFileName))
+            using (var compStream = new GZipStream(destFile, CompressionMode.Compress))
+            {
+                int singleByte = sourceFile.ReadByte();
+                while (singleByte != -1)
+                {
+                    compStream.WriteByte((byte)singleByte);
+                    singleByte = sourceFile.ReadByte();
+                }
+            }
+        }
+
+        /// <summary>Decompresses a file which was compressed using GZip compression.</summary>
+        /// <param name="inFileName">File name of compressed file.</param>
+        /// <param name="outFileName">File name to give to decompressed file.</param>
+        public void GZipDecompressFile(string inFileName, string outFileName)
+        {
+            using (FileStream sourceFile = File.OpenRead(inFileName))
+            using (FileStream destFile = File.Create(outFileName))
+            using (var compStream = new GZipStream(sourceFile, CompressionMode.Decompress))
+            {
+                int singleByte = compStream.ReadByte();
+                while (singleByte != -1)
+                {
+                    destFile.WriteByte((byte)singleByte);
+                    singleByte = compStream.ReadByte();
+                }
+            }
+        }
+    }
+}