+++ src/Common/CHANGELOG.md
@@ -16,6 +16,7 @@ Unless otherwise noted, all changes made by [@jeremyknight-me](https://github.co
 - Added GetAttribute<T> to EnumHelper.
 - Added new helper class ExcelHelper with GetColumnName(int columnNumber). 
 - Added generic ADO classes for reference purposes (migrated from older code). 
+- Added Gzip helper class with CompressFile and DecompressFile methods. 
 
 ### Changed